pyproject.toml
[project]
name = "jenosize-content-generator"
version = "0.1.0"
description = "Multi-agent content generation system for Jenosize"
authors = [
    {name = "Jenosize Team", email = "team@jenosize.com"}
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "gradio>=4.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "aiofiles>=23.2.1",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "numpy>=1.24.0,<2.0.0",
    "pandas>=2.1.0",
    "jinja2>=3.1.0",
    "loguru>=0.7.0",
    "scikit-learn>=1.3.0,<1.4.0",
    "openpyxl>=3.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]